clean up all new admin components
