+++ client/src/components/admin/searchResults/ResultsCount.jsx
